set up app routes using Routes instead of createBrowserRouter
