+++ client/.gitignore
@@ -25,3 +25,6 @@ DS_Store
 *.njsproj
 *.sln
 *.sw?
+
+#env files
+.env
