+++ client/src/main.jsx
@@ -2,5 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "@src/App";
 import "@src/index.css";
+import { Auth0Provider } from "@auth0/auth0-react";
+import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.createRoot(document.getElementById("root")).render(<App />);
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <BrowserRouter>
+    <Auth0Provider>
+      <App />
+    </Auth0Provider>
+  </BrowserRouter>
+);
